feat: adjusting how reduz working after now, start using a little bit more functional programming and init of separating of server and client functions
